Update on the PrestaShop DevDocs about the hook on the moderns pages
Hello I want to contribute to the community so here is my proposition. I was learning the prestashop 1.7.6 with this doc but it's not up to date so the part on the hooks on moderne pages wasn't working. I had to change the namespace and to create the method getRepository to create the repository and load the service without this, the erreur : You have requested a non-existent service is thrown