Updating context of where code goes
It's not clear that we are still working with the mymodule.php file because the code block starts with a new <?php block which tricked me into thinking that we were in a new file. I added a line to make sure it's known to work in the mymodule.php file.