Please clarify the correct way to autoload
I'm adding the steps I had to do to actually make the command accessible and working on our Prestashop 1.7.5.2 installation as I could find no other documentation on the proper procedure apart from a sample module that can be installed via a custom repository. Unfortunately that way of installing wasn't possible in our case and we had to work with an existing legacy module that needed to append a new console command for handling extensively long cron tasks. 

This is documentation purely based on experience on our installation and maybe it will help others looking to do the same. It doesn't mean it's the proper way, it's the only way I could find by experience to make things work. 

If there's a suggested way of doing it, please scratch this out and write it so that we can use everything correctly in the recommended way.