Update tutorial.md
I thing there's been a confusion between camelCase and PascalCase. CamelCase is a convention used in programming and computer-related fields to write compound words or phrases where each word within the compound starts with a capital letter, except for the first word. In PascalCase, also known as UpperCamelCase, each word within a compound starts with an uppercase letter, including the first word. "MyModule" is in PascalCase, in your doc, it's defined as camelCase.